Updates, fixes, and a rename
When you notice you are 2(3) of the top searches on google for
"impregnate.lua" (and not Putnam's MUCH improved version at:

http://www.bay12forums.com/smf/index.php?topic=139553.msg7036221#msg7036221

You tend to recall you HAVE a script repo and you've not touched it well
since 42.xx? Earlier? Is this the final update before I consider it 43.xx?

I dunno. But it needed updating... and I found some updates.
 Changes to be committed:
	modified:   adaptation-eld.rb

Just my version.

	modified:   civviesplode.lua

Updated.

	modified:   civviesplosion.lua

Updated

	modified:   conscript-eld.lua

Very Updated.

	renamed:    create-items.rb -> create-items2.rb

Not sure my tweaked version is up to date any more... no time to test at
the moment.

	modified:   exportlegendsWV-eld.lua

Updated... but it should be unneeded now.

	new file:   fix/growthbug.rb

I still use it... like many of the files here that likely are no longer
part of DFHack's script collection.

	modified:   impregnate.lua

Still works... but it's from the forums from 2014.